reorganize device sync
